locate ensure_pragman_key more conveniently
